refactor: reorganiza imports e cria componentes de formatação para CPF, data e status
