+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/CoeqDataTable/index.tsx
@@ -26,7 +26,7 @@ import { coeqColumns } from "./consts";
 import * as service from "./service";
 import { FiltersContext } from "@/features/acf/frontend/common/WithFilters/context";
 import type { AppliedFiltersCoeq } from "../../model";
-import { EmptyTableMessage } from "../../common/EmptyTableMessage";
+import { EmptyTableMessage } from "../common/EmptyTableMessage";
 
 export type { AppliedFiltersCoaps, AppliedFiltersCoeq } from "../../model";
