+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/common/NameFormatter/__tests__/NameFormatter.test.tsx
